requirements.txt
pytest
codecov
pytest-cov
mypy
sphinx
numpy
scipy
wheel
flake8
pandas
pillow>=7.0.0
tqdm
scikit-learn
